Fixed tab display problem for users with the SuperTab Vim plugin
When VimDebug pastes output from the debugger in the Vim scratch buffer,
SuperTab interprets tabs as coming from the user and tries to do tab
completion.  To prevent this, the code now places debugger output in a
register and pastes that to the scratch buffer.  The register's old value is
restored.