Allow users to use a non-system perl.
Thanks to '/usr/bin/env perl', their PATH will be searched, as if they
had called the program from the command line with "perl ...".

With this mechanism, the '-w' flag cannot be used on the shebang line,
but anyway it is probably better to have 'use warnings;', which can be
scoped, contrary to '-w', which is global.