Fixed an edge case.
The code should now corectly parse out the file name and line number
from lines like:

    Class::MOP::Method::Generated::CODE(0xa1e4988)(accessor amount defined at lib/Currency.pm:8):

Note that if a file or directory contains a space, the code will break.
Note also that in that particular case we have a relative path, and
also, it appears to be generated code; I'm not sure how the rest of the
code will handle that.