Fixed a bug that prevented the end of the debugged program from being detected.
The recent improvements to parse out the file name and line number from
the debugger output changed the original way of doing it in that if no
file name or line number was found, it would not touch the current
setting, leaving it usually defined, while the original would set it to
undef.  Another part of the program(act() in vdd) checked the
definedness of these settings and as a result of their now being pretty
much always defined, took a wrong decision, causing the problem.