Better quit behaviour.
It tries hard not to leave any fifos or debugger console data lying
around.

If <F12> is pressed when the debugger is already running, it sets the
focus to the debugger console.

The implementation now uses a real file instead of a Vim scratch buffer.
This allows us to trap launch errors.