Switched from using a fifo to using sockets
This change is a step towards making VimDebug work on Windows.  POSIX::mkfifo
doesn't work on Windows.  To add support for Windows, I would have needed to
use Win32::Pipe and that means a separate chunk of code just for Windows.  But
IO::Socket should work on all platforms.  So probably this will keep the code
simpler.