Debug Perl only from now on.
Since we have no support for languages other than Perl for now (and for
the foreseeable future), it's always the perl debugger that will be
invoked from now on. And it will be invoked for the file in the current
Vim window no matter what its filename or extension. Also, arguments
will be prompted for in all cases.

Note that the backend modules are intact, and if anyone ever proposes
producing a role file like VimDebug::Ruby for example, we can think of
fixing the interface accordingly.