Have a gvim menu for VimDebug.
Actually, there are many changes in this commit, and a lot of
refactoring. For example, the port number (6543) is now supplied by the
VimDebug::Daemon->port() method, and the "done file" ('.vdd.done'), by
the VimDebug::Protocol->k_doneFile() method.

But the biggest changes are in the VimDebug.vim file:

User configurable variable definitions are grouped together at the
beginning of the file, and are named like "s:cfg_fooBar". Similarly,
global variables whose values may have been set before loading the
VimDebug module are grouped there also.

There is now a 's:Initialize' function used to, yes, initialize
everything necessary for proper functioning of VimDebug. It is defined
early in the file, and it is called at the end,

There are now Vim commands (as opposed to functions) only to launch the
debugger and to toggle the key bindings (VDstart, VDtoggleKeys).

A good half-dozen functions were added to manage the interface settings
(key bindings, menus).